Level A-Packages, organised into packages
