+++ src/main/java/duke/Event.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.DateTimeInterface;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
