+++ src/main/java/duke/MarkCommand.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.Command;
+
 public class MarkCommand extends Command {
     private final int index;
